Extracted test_suite_detail_tabs.py
